Allow cache though enabled repositories to fetch image from upstream
Allow read only access "users" to use pull through cache feature for new images in the repository.

why:
We are using ecr-public pull through cache and we want also new images to be downloaded
automatically to the cache. Therefore read only access users need to be able to do that.
There is currently no support for the cache feature in this module and I wanted to keep
the changes minimal.